Fixed bugs and organized code
Still a bit of a mess but I fixed a few of the major bugs.
Sending commands to the device is much smoother now (there
is a possible bug in USBHwEPRead(), see IMU.c) and the hostside
program no longer quits when a stop is issued.

imu-device-host-interface holds macros and definitions used in both
device and hostside code.